event-bot/requirements.txt
discord
python-dotenv
autopep8
requests